+++ docs/CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 ## [Unreleased]
 
+## [0.3.0] - 2024-07-10
+
 ### Added
 
 -   Added validation for trove classifiers
@@ -55,7 +57,8 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
-[Unreleased]: https://github.com/spapanik/phosphorus/compare/v0.2.1...main
+[Unreleased]: https://github.com/spapanik/phosphorus/compare/v0.3.0...main
+[0.3.0]: https://github.com/spapanik/phosphorus/compare/v0.2.1...v0.3.0
 [0.2.1]: https://github.com/spapanik/phosphorus/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/spapanik/phosphorus/compare/v0.1.3...v0.2.0
 [0.1.3]: https://github.com/spapanik/eulertools/compare/v0.1.2...v0.1.3